Automatically check any outstanding deployments on every command. Does incur some lag, so leaving this as a separate, optional commit.
