+++ src/command_modules/azure-cli-ml/azure/cli/command_modules/ml/_util.py
@@ -79,6 +79,66 @@ class CommandLineInterfaceContext(object):
             if outer_match_obj:
                 self.hdi_home = outer_match_obj.group('cluster_name')
         self.hdi_domain = self.hdi_home.split('.')[0] if self.hdi_home else None
+        self.check_deployments(silent=True)
+
+    def check_deployments(self, silent=False):
+        from ._az_util import az_check_template_deployment_status
+        from ._az_util import AzureCliError
+        deployments = self.get_deployments()
+        if not deployments:
+            if not silent:
+                print('No active deployments found.')
+                print('To begin provisioning an environment, run')
+                print('  az ml env setup [-k]')
+            return
+        for deployment_id in deployments:
+            if not silent:
+                print('Checking on deployment {}...'.format(deployment_id))
+            try:
+                completed_deployment = az_check_template_deployment_status(deployment_id, silent)
+            except AzureCliError as exc:
+                print(exc.message)
+                return
+
+            if completed_deployment:
+                try:
+                    acs_master = completed_deployment.properties.outputs['masterFQDN'][
+                        'value']
+                    acs_agent = \
+                    completed_deployment.properties.outputs['agentpublicFQDN']['value']
+                    if acs_master and acs_agent:
+                        print('ACS deployment {} has completed.'.format(deployment_id))
+                        print('ACS Master URL     : {}'.format(acs_master))
+                        print('ACS Agent URL      : {}'.format(acs_agent))
+                        print(
+                            'ACS admin username : acsadmin (Needed to set up port forwarding in cluster mode).')
+                        print(
+                            'To configure az ml with this environment, set the following environment variables.')
+                        print('{} has also been updated.'.format(
+                            os.path.join(os.path.expanduser('~'), '.amlenvrc')))
+                        if platform.system() in ['Linux', 'linux', 'Unix', 'unix']:
+                            write_acs_to_amlenvrc(acs_master, acs_agent, "export")
+                        else:
+                            write_acs_to_amlenvrc(acs_master, acs_agent, "set")
+
+                        try:
+                            ssh_config_fp = os.path.join(os.path.expanduser('~'), '.ssh',
+                                                         'config')
+                            with open(ssh_config_fp, 'a+') as sshconf:
+                                sshconf.write('Host {}\n'.format(acs_master))
+                                sshconf.write('    HostName {}\n'.format(acs_master))
+                                sshconf.write('    User acsadmin\n')
+                                sshconf.write('    IdentityFile ~/.ssh/acs_id_rsa\n')
+                            os.chmod(ssh_config_fp, 0o600)
+                        except:
+                            print('Failed to update ~/.ssh/config. '
+                                  'You will need to manually update your '
+                                  '.ssh/config to look for ~/.ssh/acs_id_rsa '
+                                  'for host {}'.format(acs_master))
+                        self.remove_deployment(deployment_id)
+                        print("To switch to cluster mode, run 'az ml env cluster'.")
+                except AzureCliError as exc:
+                    print(exc.message)
 
 
     @staticmethod
@@ -326,6 +386,20 @@ class JupyterContext(CommandLineInterfaceContext):
 
 
 # UTILITY FUNCTIONS
+def write_acs_to_amlenvrc(acs_master, acs_agent, env_verb):
+    env_statements = ["{} AML_ACS_MASTER={}".format(env_verb, acs_master),
+                      "{} AML_ACS_AGENT={}".format(env_verb, acs_agent)]
+
+    print('\n'.join([' {}'.format(statement) for statement in env_statements]))
+    try:
+        with open(os.path.expanduser('~/.amlenvrc'), 'a+') as env_file:
+            env_file.write('\n'.join(env_statements) + '\n')
+    except IOError:
+        pass
+
+    print('')
+
+
 def get_json(payload):
     """
     Handles decoding JSON to python objects in py2, py3
